+++ CHANGES.rst
@@ -15,6 +15,14 @@ Changelog
 - Plone 5.1 compatibility
   [alecm]
 
+- Port to plone.app.testing
+  [tomgross]
+
+- Remove unused egg dependencies
+  [tomgross]
+
+- implements -> implementer
+  [tomgross]
 
 1.6.0 (2017-02-21)
 ------------------